moved the contents os_bsp.c to hal_bsp.c for the pic32 based BSPs to fit in with the rest of the boards
